sure: An R package for constructing surrogate-based residuals and diagnostics for ordinal and general regression models.

The sure package provides surrogate-based residuals for fitted ordinal and general (e.g., binary) regression models of class ordinal::clm(), stats::glm(), rms::lrm(), rms::orm(), MASS::polr(), or VGAM::vglm().

Details

The development version can be found on GitHub: https://github.com/AFIT-R/sure. As of right now, sure exports the following functions:

resids - construct (surrogate-based) residuals;

plot - plot diagnostics using tinyplot::tinyplot()-based graphics;

gof - simulate p-values from a goodness-of-fit test.

References

Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal Regression Models: A Surrogate Approach. Journal of the American Statistical Association (accepted).

See also

Author

Maintainer: Brandon Greenwell greenwell.brandon@gmail.com

Authors:

Andrew McCarthy mccarthyan@gmail.com

Boehmke Bradley bradleyboehmke@gmail.com

Other contributors:

Liz Glaser liz.glaser@gmail.com [contributor]